Allow ModCandidate.copyToFile to replace existing files, needed becuase the RMR creates an empty temp file. Fix RuntimeModRemapper error handling

+++ src/main/java/net/fabricmc/loader/impl/discovery/RuntimeModRemapper.java
@@ -158,6 +158,10 @@ public final class RuntimeModRemapper {
 			remapper.finish();
 
 			for (RemapInfo info : infoMap.values()) {
+				if (info.outputPath == null) {
+					continue;
+				}
+
 				try {
 					Files.deleteIfExists(info.outputPath);
 				} catch (IOException e) {
